Reduce the public API surface area of the types to just the selectors, PerPackageEntities, and EntityRecord
